requirements.txt
Django>=4.2,<5.0
langchain>=0.1.0
requests
python-dotenv
